Unify both area problems into one main file
