+++ examples/petsc/qfunctions/area/areacube.okl
@@ -52,7 +52,7 @@ typedef double CeedScalar;
 // *****************************************************************************
 
 // -----------------------------------------------------------------------------
-@kernel void SetupMassGeo(void *ctx, CeedInt Q,
+@kernel void SetupMassGeoCube(void *ctx, CeedInt Q,
                           const int *iOf7, const int *oOf7,
                           const CeedScalar *in, CeedScalar *out) {
 // Inputs
